# app/Http/Controllers/API/UserController.php
<?php

namespace App\Http\Controllers\API;

use App\Http\Controllers\Controller;
use App\Models\User;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Hash;
use Illuminate\Validation\ValidationException;
use OpenApi\Attributes as OA;

class UserController extends Controller
{
    #[OA\Post(
        path: '/api/register',
        summary: "Inscription",
        tags: ["Authentification"],
        parameters: [new OA\Parameter(name: 'Accept', in: 'header', example: 'application/json')],
        requestBody: new OA\RequestBody(content: new OA\JsonContent(example: ['name' => 'Svet', 'email' => 'svet@mail.com', 'password' => 'password123'])),
        responses: [
            new OA\Response(response: 201, description: 'Créé'),
            new OA\Response(response: 422, description: 'Erreur validation')
        ]
    )]
    public function register(Request $request)
    {
        $validated = $request->validate([
            'name' => 'required|string|max:255',
            'email' => 'required|string|email|max:255|unique:users',
            'password' => 'required|string|min:8',
        ]);

        $user = User::create([
            'name' => $validated['name'],
            'email' => $validated['email'],
            'password' => Hash::make($validated['password']),
        ]);

        return response()->json([
            'message' => 'User Created',
            'user' => $user
        ], 201);
    }

    #[OA\Post(
        path: '/api/login',
        summary: "Connexion",
        tags: ["Authentification"],
        parameters: [new OA\Parameter(name: 'Accept', in: 'header', example: 'application/json')],
        requestBody: new OA\RequestBody(content: new OA\JsonContent(example: ['email' => 'svet@mail.com', 'password' => 'password123'])),
        responses: [
            new OA\Response(response: 200, description: 'Succès'),
            new OA\Response(response: 401, description: 'Identifiants incorrects'),
            new OA\Response(response: 422, description: 'Erreur validation')
        ]
    )]
    public function login(Request $request)
    {
        $request->validate([
            'email' => 'required|string|email',
            'password' => 'required|string',
        ]);

        $user = User::firstWhere('email', $request->email);

        if (! $user || ! Hash::check($request->password, $user->password)) {
            return response()->json(['message' => 'Incorrect password or email.'], 401);
        }

        $token = $user->createToken('api-token')->plainTextToken;

        return response()->json([
            'message' => 'Success',
            'token' => $token,
            'user' => $user
        ]);
    }

    #[OA\Post(
        path: '/api/logout',
        summary: "Déconnexion",
        security: [['bearerAuth' => []]],
        tags: ["Authentification"],
        parameters: [new OA\Parameter(name: 'Accept', in: 'header', example: 'application/json')],
        responses: [
            new OA\Response(response: 200, description: 'Succès'),
            new OA\Response(response: 401, description: 'Non autorisé')
        ]
    )]
    public function logout(Request $request)
    {
        $request->user()->currentAccessToken()->delete();

        return response()->json(['message' => 'User Disconnected']);
    }
}